# Refactor threadpool's inner workings