--- a/src/join.rs
+++ b/src/join.rs
@@ -1,9 +1,9 @@
-use latch::{LockLatch, SpinLatch};
+use latch::{Latch, LockLatch, SpinLatch};
 #[allow(unused_imports)]
 use log::Event::*;
 use job::StackJob;
 use thread_pool::{self, WorkerThread};
-use std::mem;
+use std::any::Any;
 use unwind;
 
 /// The `join` function takes two closures and *potentially* runs them
@@ -19,54 +19,57 @@ pub fn join<A, B, RA, RB>(oper_a: A, oper_b: B) -> (RA, RB)
     unsafe {
         let worker_thread = WorkerThread::current();
 
-        // slow path: not yet in the thread pool
+        // Slow path: not yet in the thread pool.
         if worker_thread.is_null() {
             return join_inject(oper_a, oper_b);
         }
 
-        log!(Join { worker: (*worker_thread).index() });
+        let worker_thread = &*worker_thread;
 
-        // create virtual wrapper for task b; this all has to be
+        log!(Join { worker: worker_thread.index() });
+
+        // Create virtual wrapper for task b; this all has to be
         // done here so that the stack frame can keep it all live
-        // long enough
+        // long enough.
         let job_b = StackJob::new(oper_b, SpinLatch::new());
-        (*worker_thread).push(job_b.as_job_ref());
+        let job_b_ref = job_b.as_job_ref();
+        worker_thread.push(job_b_ref);
 
-        // record how many async spawns have occurred on this thread
-        // before task A is executed
-        let spawn_count = (*worker_thread).current_spawn_count();
+        // Execute task a; hopefully b gets stolen in the meantime.
+        let result_a = match unwind::halt_unwinding(oper_a) {
+            Ok(v) => v,
+            Err(err) => join_recover_from_panic(worker_thread, &job_b.latch, err),
+        };
 
-        // execute task a; hopefully b gets stolen
-        let result_a;
-        {
-            let guard = unwind::finally(&job_b.latch, |job_b_latch| {
-                // If another thread stole our job when we panic, we must halt unwinding
-                // until that thread is finished using it.
-                if (*WorkerThread::current()).pop().is_none() {
-                    job_b_latch.spin();
+        // Now that task A has finished, try to pop job B from the
+        // local stack.  It may already have been popped by job A; it
+        // may also have been stolen. There may also be some tasks
+        // pushed on top of it in the stack, and we will have to pop
+        // those off to get to it.
+        while !job_b.latch.probe() {
+            if let Some(job) = worker_thread.pop() {
+                if job == job_b_ref {
+                    // Found it! Let's run it.
+                    //
+                    // Note that this could panic, but it's ok if we unwind here.
+                    log!(PoppedRhs { worker: worker_thread.index() });
+                    let result_b = job_b.run_inline();
+                    return (result_a, result_b);
+                } else {
+                    log!(PoppedJob { worker: worker_thread.index() });
+                    worker_thread.execute(job);
                 }
-            });
-            result_a = oper_a();
-            mem::forget(guard);
+            } else {
+                // Local deque is empty. Time to steal from other
+                // threads.
+                log!(LostJob { worker: worker_thread.index() });
+                worker_thread.wait_until(&job_b.latch);
+                debug_assert!(job_b.latch.probe());
+                break;
+            }
         }
 
-        // before we can try to pop b, we have to first pop off any async spawns
-        // that have occurred on this thread
-        (*worker_thread).pop_spawned_jobs(spawn_count);
-
-        // if b was not stolen, do it ourselves, else wait for the thief to finish
-        let result_b;
-        if (*worker_thread).pop().is_some() {
-            log!(PoppedJob { worker: (*worker_thread).index() });
-            result_b = job_b.run_inline(); // not stolen, let's do it!
-        } else {
-            log!(LostJob { worker: (*worker_thread).index() });
-            (*worker_thread).steal_until(&job_b.latch); // stolen, wait for them to finish
-            result_b = job_b.into_result();
-        }
-
-        // now result_b should be initialized
-        (result_a, result_b)
+        return (result_a, job_b.into_result());
     }
 }
 
@@ -86,4 +89,17 @@ unsafe fn join_inject<A, B, RA, RB>(oper_a: A, oper_b: B) -> (RA, RB)
     job_b.latch.wait();
 
     (job_a.into_result(), job_b.into_result())
+}
+
+/// If job A panics, we still cannot return until we are sure that job
+/// B is complete. This is because it may contain references into the
+/// enclosing stack frame(s).
+#[cold] // cold path
+unsafe fn join_recover_from_panic(worker_thread: &WorkerThread,
+                                  job_b_latch: &SpinLatch,
+                                  err: Box<Any + Send>)
+                                  -> !
+{
+    worker_thread.wait_until(job_b_latch);
+    unwind::resume_unwinding(err)
 }