This is groundwork for better handling of asynchronous tasks and futures integration. 

There are a few changes:

- We no longer track the local stack depth. Instead, we pop and use pointer comparisons to decide if we popped the right-hand-side task. This scales gracefully to unbalanced pushes/pops (which can occur with `spawn_async` and with blocking).
- The main loop is now using the same code as the steal loop.
- Individual worker threads will now go to sleep when stealing if new data doesn't seem to be showing up. This is somewhat tunable: you can configure thresholds of how long they will yield before attempting to sleep.

I did some benchmarking throughout. I should do more though as I've done various refactorings and cleanup to the core algorithms in preparing this PR. In any case, [here is a plot](https://people-mozilla.org/~nmatsakis/spindown.svg) charting the effect of these changes on two join microbenchmarks that I added, since I wanted to ensure that this core underlying API remained solid. The plot shows five runs of various key commits: the most important commit is the final one, which is the effect of all the commits together. (The hash doesn't *quite* match because I dropped an inconsequential commit in the process.) The first two are basically a baseline. (This chart was created with the [cargo-chrono tool](https://github.com/nikomatsakis/cargo-chrono) that I implemented just for this purpose...)

Some things to do before landing:

- [x] Double check on panic tests and panic propagation.
- [x] Some kind of termination test cases -- do we even have any?
- [x] Write up good docs for the sleep module.
- [x] Tighten up some of the atomic orderings.
- [x] Gather more benchmark data.