@nikomatsakis 2016-12-02
I did more measurements using a variety of benchmarks. Mostly they tell the same story: overall this change is a bit faster. The one exception is quick-sort.

Here is a plot of running a select set of tests, 10 times each, for various notable commits:

https://people-mozilla.org/~nmatsakis/2016.12.02-rayon-pr172.svg

This is the same data, but using medians of the 10 runs (with error bars):

https://people-mozilla.org/~nmatsakis/2016.12.02-rayon-pr172-medians.svg

Finally, I re-ran the results with quick-sort and found that indeed the same shift appears:

https://people-mozilla.org/~nmatsakis/2016.12.02-rayon-pr172-quicksort-medians.svg