@nikomatsakis 2016-12-03
Ok, found a rather large oversight -- somewhere in my rebase I lost a critical call to `tickle()` that would cause the workers to wake up when other workers pushed new jobs. Amazing that this didn't have a more dramatic effect on performance, really. I am now re-measuring the data in light of fixing this problem -- but at least for quicksort, the results are encouraging (the perf is back to where it was).