@nikomatsakis 2016-12-03
Revised data plots look pretty good:

- [all points](https://people-mozilla.org/~nmatsakis/2016.12.03-rayon-pr172-all.svg)
- [medians](https://people-mozilla.org/~nmatsakis/2016.12.03-rayon-pr172-medians.svg)