@nikomatsakis 2016-12-03
Running mergesort with more data doesn't seem to show any performance loss:

```
# this branch, two runs:
seq: sorted 250000000 ints: 27.862516 s
par: sorted 250000000 ints: 2.594257 s
speedup: 10.74x

seq: sorted 250000000 ints: 27.71634 s
par: sorted 250000000 ints: 2.6546683 s
speedup: 10.44x

# master, two runs:
seq: sorted 250000000 ints: 27.862516 s
par: sorted 250000000 ints: 2.594257 s
speedup: 10.74x

seq: sorted 250000000 ints: 27.71634 s
par: sorted 250000000 ints: 2.6546683 s
speedup: 10.44x
```