@nikomatsakis 2016-12-04
I have tweaked the memory orderings and written up my thoughts on why they are what they are. I also did measurements with everything set to `SeqCst` and saw no particular impact on performance -- so, in the `sleep` module at least, it may be worth forgoing complex reasoning and just using `SeqCst` everywhere. But it seems good to think rigorously and try for better `Ordering` values where possible.