@nikomatsakis 2016-12-20
Rebased. 

@emilio I'm curious if you can easily run some servo benchmarks here to see if there is any effect from this branch. I'd like to land this soon, in any case.