@emilio 2016-12-21
I think I can, but not in the really short-term (I need to do a bunch of classwork this week, though I'll let you know if I manage to get some spare time to do this :P).

FWIW what I use to do is downloading a local copy of a big DOM tree (the [html single-page spec](https://html.spec.whatwg.org/) uses to work fine), and then profile Servo loading that page locally (network makes stuff less deterministic).

To profile you can use Servo's `-p <period>` option, though in my experience it's not really reliable, so I always end up falling back to `perf`. When measuring rayon's performance, you usually care about the `top_down_dom` function, or related ones in the style system.