@nikomatsakis 2016-12-23
OK, I've settled on how I think we should handle the atomic orderings and written some docs. I also rebased and cleaned up the history. The main "to do" item now is to add some termination and panic propagation test cases. I changed some of the panic recovery behavior in this branch -- I would prefer to say "normalized" -- and we ought to test that. (For example, if you call `join(a, b)` and `a` panics, it used to be nondeterminstic whether `b` would execute. This is no longer the case: `b` will always execute.)