@nikomatsakis 2016-12-23
I now added several tests around panic propagation and tried to document the behavior with some precision. =)