@nikomatsakis 2016-12-23
OK, I added some thread-pool tests and did some final cleanup. I think this branch is ready to land.