@nikomatsakis 2016-12-23
Here is [a plot of measurements for these changes][plot], showing before/after with the final rebased version, normalized to 100:

![plot][plot]

plot: https://people-mozilla.org/~nmatsakis/rayon-pr172.svg

There are only two regressions (the colors aren't working right, so you have to hover the mouse):

- merge-sort, ~8%
- parallel-find-last, ~5%

When I measured the merge-sort with a larger input, I found the regression disappeared. I can't really account for the 8% we see here in the `#[bench]` results. I'm also not sure why `parallel-find-missing` seems to be so much faster.