@nikomatsakis 2016-12-23
Re-ran the measurements. Looks the same. Some charts:

- https://people-mozilla.org/~nmatsakis/rayon-pr172-1.svg (normalized medians)
- https://people-mozilla.org/~nmatsakis/rayon-pr172-1-points.svg (all data points)